refactor: move client exceptions to api module
