+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/smarthometech/commerce/api/exception/InsufficientQuantityException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.smarthometech.commerce.warehouse.application;
+package ru.yandex.practicum.smarthometech.commerce.api.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
