+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/smarthometech/commerce/api/exception/WarehouseInteractionException.java
@@ -1,0 +1,12 @@
+package ru.yandex.practicum.smarthometech.commerce.api.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.INTERNAL_SERVER_ERROR)
+public class WarehouseInteractionException extends RuntimeException {
+
+    public WarehouseInteractionException(String message) {
+        super(message);
+    }
+}
