+++ commerce/warehouse/src/main/java/ru/yandex/practicum/smarthometech/commerce/warehouse/application/WarehouseService.java
@@ -15,6 +15,9 @@ import ru.yandex.practicum.smarthometech.commerce.api.dto.warehouse.AddressDto;
 import ru.yandex.practicum.smarthometech.commerce.api.dto.warehouse.BookedProductsDto;
 import ru.yandex.practicum.smarthometech.commerce.api.dto.warehouse.NewProductInWarehouseRequest;
 import ru.yandex.practicum.smarthometech.commerce.api.dto.warehouse.ShoppingCartDto;
+import ru.yandex.practicum.smarthometech.commerce.api.exception.InsufficientQuantityException;
+import ru.yandex.practicum.smarthometech.commerce.api.exception.ProductAlreadyExistsException;
+import ru.yandex.practicum.smarthometech.commerce.api.exception.ProductNotFoundException;
 import ru.yandex.practicum.smarthometech.commerce.warehouse.domain.WarehouseItem;
 import ru.yandex.practicum.smarthometech.commerce.warehouse.domain.WarehouseRepository;
